% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{calc_mgidi}
\alias{calc_mgidi}
\title{helpers}
\usage{
calc_mgidi(data, rhot_table = NULL, SI = 15, mineval = 1, avg_NA = FALSE)
}
\arguments{
\item{data}{data frame, with "genotype" as column names, and variables defined in `rhot_table`}

\item{rhot_table}{a data frame with columns "trait", "direction", "opti_val" and "weight". Direction must be one of "min", "max" or "opti". In that case, the "opti_val" column should be filled for this line.}

\item{SI}{selection intensity in percentage of number of unique genotypes in the table. Default to 15, must be between 0 and 100.}

\item{mineval}{parameter of `mgidi` function for the number of factors to retain. Default to 1.}

\item{avg_NA}{logical to indicate if missing data would be replaced by population average to avoid the exclusion of genotypes in MGIDI.}
}
\value{
a list with two elements: "res_mgidi" the MGIDI output and "data_mean", the averaged data per genotype for all selected variables.
}
\description{
This function computes the MGIDI algorithm to select genotypes based on a selection index.
}
